Method for killing child processes for Windows was corrected

+++ src/core/lib/testHelpers.js
@@ -90,14 +90,22 @@ export default class {
         const cmd = commandArr.shift();
         const execObj = spawn(cmd, commandArr, {
             cwd: path.resolve(__dirname, "../../.."),
-            detached: true,
+            // detached: true,
         });
         return execObj;
     }
 
     /* istanbul ignore file */
-    killProcess(pid) {
-        process.kill(-pid);
+    killProcess(childProcess, pid) {
+        if (childProcess) {
+            childProcess.stdout.destroy();
+            childProcess.stderr.destroy();
+            childProcess.kill();
+            // process.kill(-(childProcess.pid));
+        }
+        if (pid) {
+            process.kill(-pid);
+        }
     }
 
     /* istanbul ignore file */
